refactor(daengle): 쿼리 interface 경로 변경
